Organized packages in test folder
