+++ src/main/kotlin/com/montebruni/holder/wallet/application/client/CustomerClient.kt
@@ -1,0 +1,9 @@
+package com.montebruni.holder.wallet.application.client
+
+import com.montebruni.holder.wallet.application.client.response.CustomerResponse
+import java.util.UUID
+
+interface CustomerClient {
+
+    fun findById(id: UUID): CustomerResponse?
+}
