+++ src/main/kotlin/com/montebruni/holder/wallet/application/client/exception/CustomerNotFoundException.kt
@@ -1,0 +1,3 @@
+package com.montebruni.holder.wallet.application.client.exception
+
+class CustomerNotFoundException : Exception("Customer not found")
