+++ src/main/kotlin/com/montebruni/holder/wallet/application/client/response/CustomerResponse.kt
@@ -1,0 +1,7 @@
+package com.montebruni.holder.wallet.application.client.response
+
+import java.util.UUID
+
+data class CustomerResponse(
+    val id: UUID
+)
