--- src/main/kotlin/com/montebruni/holder/wallet/application/dataprovider/CustomerDataProvider.kt
@@ -1,9 +1,0 @@
-package com.montebruni.holder.wallet.application.dataprovider
-
-import com.montebruni.holder.wallet.application.dataprovider.data.CustomerData
-import java.util.UUID
-
-interface CustomerDataProvider {
-
-    fun findById(id: UUID): CustomerData?
-}
