--- src/main/kotlin/com/montebruni/holder/wallet/application/dataprovider/data/CustomerData.kt
@@ -1,7 +1,0 @@
-package com.montebruni.holder.wallet.application.dataprovider.data
-
-import java.util.UUID
-
-data class CustomerData(
-    val id: UUID
-)
