--- src/main/kotlin/com/montebruni/holder/wallet/application/dataprovider/exception/CustomerDataProviderNotFoundException.kt
@@ -1,3 +1,0 @@
-package com.montebruni.holder.wallet.application.dataprovider.exception
-
-class CustomerDataProviderNotFoundException : Exception("Customer not found")
