+++ src/main/kotlin/com/montebruni/holder/wallet/application/usecase/impl/CreateWalletImpl.kt
@@ -1,7 +1,7 @@
 package com.montebruni.holder.wallet.application.usecase.impl
 
-import com.montebruni.holder.wallet.application.dataprovider.CustomerDataProvider
-import com.montebruni.holder.wallet.application.dataprovider.exception.CustomerDataProviderNotFoundException
+import com.montebruni.holder.wallet.application.client.CustomerClient
+import com.montebruni.holder.wallet.application.client.exception.CustomerNotFoundException
 import com.montebruni.holder.wallet.application.event.EventPublisher
 import com.montebruni.holder.wallet.application.event.events.WalletCreatedEvent
 import com.montebruni.holder.wallet.application.usecase.CreateWallet
@@ -17,14 +17,14 @@ import kotlin.let
 @Service
 class CreateWalletImpl(
     private val walletRepository: WalletRepository,
-    private val customerDataProvider: CustomerDataProvider,
+    private val customerClient: CustomerClient,
     private val eventPublisher: EventPublisher
 ) : CreateWallet {
 
     override fun execute(input: CreateWalletInput): CreateWalletOutput {
-        customerDataProvider.findById(input.customerId) ?: throw CustomerDataProviderNotFoundException()
+        customerClient.findById(input.customerId) ?: throw CustomerNotFoundException()
         input.managerId?.let {
-            customerDataProvider.findById(input.managerId) ?: throw CustomerDataProviderNotFoundException()
+            customerClient.findById(input.managerId) ?: throw CustomerNotFoundException()
         }
 
         return input
