--- src/test/kotlin/com/montebruni/holder/account/application/domain/email/UserCreatedMailTemplateTest.kt
@@ -1,17 +1,0 @@
-package com.montebruni.holder.account.application.domain.email
-
-import org.junit.jupiter.api.Assertions.assertEquals
-import org.junit.jupiter.api.Test
-
-class UserCreatedMailTemplateTest {
-
-    @Test
-    fun `should create mail template with password`() {
-        val password = "password"
-        val template = UserCreatedMailTemplate(password)
-
-        val body = template.getBody()
-
-        assertEquals("Your account has been created successfully. Your new password is $password.", body)
-    }
-}
